make sure to delete any kids that aren't in the system, from the admin_auths table - cron job

+++ mashpia.com/cron/syncDeleted.php
@@ -3,5 +3,5 @@
 include_once( __DIR__ . '/public/api/header/db.php' );
 
 // create cron job to delete admin_auth associations that point to missing child
-$sql = "delete from admin_auths where id not in (select user_id from users)";
+$sql = "delete from admin_auths where auth = 'user' and id not in (select user_id from users)";
 echo mysql_query($sql);
